requirements.txt
aiohttp==3.6.3
async-timeout==3.0.1
attrs==20.3.0
autopep8==1.5.4
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
discord.py==1.5.1
idna==2.10
multidict==4.7.6
pycodestyle==2.6.0
pycparser==2.20
PyNaCl==1.3.0
requests==2.25.0
six==1.15.0
spotipy==2.16.1
toml==0.10.2
urllib3==1.26.2
wavelink==0.9.6
wincertstore==0.2
yarl==1.5.1
youtube-dl==2020.11.29
youtube-search-python==1.3.2
